Refactored maven testing setup and added ability to run parallel tests

+++ janusgraph-test/src/main/java/org/janusgraph/diskstorage/IDAuthorityTest.java
@@ -118,7 +118,7 @@ public abstract class IDAuthorityTest {
         //hasFixedUid = !config.get(IDAUTHORITY_RANDOMIZE_UNIQUEID);
         hasFixedUid = !ConflictAvoidanceMode.GLOBAL_AUTO.equals(config.get(IDAUTHORITY_CONFLICT_AVOIDANCE));
         hasEmptyUid = uidBitWidth==0;
-        blockSize = config.get(IDS_BLOCK_SIZE);
+        blockSize = config.get(IDS_BLOCK_SIZE).intValue();
         idUpperBoundBitWidth = 30;
         idUpperBound = 1l<<idUpperBoundBitWidth;
     }
